+++ src/Plugin.php
@@ -54,7 +54,7 @@ class Plugin extends BasePlugin
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
     public bool $hasReadOnlyCpSettings = true;
-    public string $schemaVersion = '1.0.0.7';
+    public string $schemaVersion = '1.1.0';
 
     public function init(): void
     {